Class

+++ Class/Class.Binary/Read.cs
@@ -258,9 +258,9 @@ public class Read : Any
         Array array;
         array = this.Arg.ImportArray;
 
-        int count;
+        long count;
         count = array.Count;
-        int i;
+        long i;
         i = 0;
         while (i < count)
         {
@@ -278,9 +278,9 @@ public class Read : Any
         Array array;
         array = this.Arg.PartArray;
 
-        int count;
+        long count;
         count = array.Count;
-        int i;
+        long i;
         i = 0;
         while (i < count)
         {
